refactor: :fire: moved common function into binary-utils.js
