+++ packages/bot-skeleton/src/services/tradeEngine/trade/Ticks.js
@@ -5,8 +5,7 @@ import { expectPositiveInteger } from '../utils/sanitize';
 import { observer as globalObserver } from '../../../utils/observer';
 import { api_base } from '../../api/api-base';
 import debounce from 'lodash.debounce';
-
-const getLast = arr => arr && (arr.length === 0 ? undefined : arr[arr.length - 1]);
+import { getLast } from '../../api/binary-utils';
 
 let tickListenerKey;
